Added the possibility to transfer vehicles in the various garages
Thanks to these modifications, I added the possibility to transfer the vehicles in the garages paying. This change is particularly useful if, for example, you have a garage in cayo perico.

Currently I have only modified the translations for English and Italian, I would love if the community would contribute in the translations!